refactor: migrate to Python 3.12+ and Pydantic v2
This commit covers the full monorepo refactoring to modernize the stack:
requires-python bumped from >=3.9 to >=3.12 in all four packages.

--- API (src/api/) ---

- Add exceptions.py: centralize all custom exception handlers
- Add lifespan async context manager to verify DB connectivity on
  startup and call engine.dispose() on graceful shutdown
- All routes: remove per-endpoint try/except wrappers (now handled by
  the global exception handlers), use Pydantic v2 model_validate()
  instead of from_orm()

--- Core (src/core/) ---

- pyproject.toml: add pydantic>=2.0,<3 and pydantic-settings>=2.0,<3
- All schemas: Pydantic v2 migration — replace class Config/orm_mode
  with ConfigDict(from_attributes=True), replace @validator with
  @field_validator, modernize field definitions
- config.py: replace python-dotenv manual loading with pydantic-settings
- redisPubSubManager.py: replace ABC base class with Protocol-based
  contracts; refactor sync and async implementations to accept injected
  host/port/db parameters instead of relying on globals
- Utilities: Type annotation cleanups and minor modernizations

--- Desktop (src/desktop/) ---

- MainWindow.py: wrap initial worker/Redis status query in a
  _refresh_worker_status() helper that catches connection failures
  gracefully; connect task_finished and task_failed signals on init
- Components and views: type annotation updates

--- Worker (src/worker/) ---

- Tasks: wrap tasks in try/finally to guarantee cnc.disconnect(),
  redis.disconnect(), db_session.close() on all exit paths
- utils/gcodeParser.py: type annotation updates

--- Root workspace ---

- Add pytest-qt, pyqt5-stubs, celery-types as dev dependencies
- Bump pytest >=7.4,<8 → >=7.4,<9

+++ src/core/core/utilities/serial.py
@@ -10,7 +10,7 @@ class SerialService:
     def get_ports(cls):
         return serial_ports.comports()
 
-    def startConnection(self, port: str, baudrate: int, timeout: int = 2) -> str:
+    def startConnection(self, port: str, baudrate: int, timeout: float = 2) -> str:
         """Closes any previous connection and starts a new one."""
         # Close any previous serial connection
         if self.interface.is_open:
